fix: season label not showing chore: refactor component names and to file
